Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm not comfortable doing this piecemeal ... I'd like to separate out the CFX_Target removal (which can be done via script/sed/etc) from the unique_ptr work which can follow.
> 
> I'm concerned of being in a state where we mismatch frees, etc.  If all the 'targets are gone, then that can't happen.  Thoughts?

Hm, does that mean we have to do this all in one giant CL? If we just get rid of the CFX_Target and allocator code, then there is a chance we'll have leaks because freeing before wasn't necessary (destroy the allocator, free the memory).

The idea with converting to unique_ptr and removing one at a time is we know that it will be cleaned up if the holder class is cleaned up. But, we don't currently know if the holder class is cleaned up or just memory free'd do we?

Patch-set: 1