Update patch set 1
Patch Set 1:

There's a small incompatibility between CFX_RetainPtr and IFX_Retainable overrides in CFDE_CSSComputedStyle for the initial value of the ref count (0 vs. 1).

In the old way, when you new up a CFDE_CSSComputedStyle, you get a retain count of 1 automatically.  Then you pass it into a retainptr, and the count will be bumped to two.

Best to yank out the dependence on IFX_Retainable and use CFX_Retainable instead, deleting the custom ref count field.  IFX_Retainable should die when you're all done with the 4 or so classes that use it here.

Patch-set: 1
Label: Code-Review=0