Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> So is there going to be a dashboard somewhere, or is this meant to be used locally/manually for now?


For now, manually. We could look into getting a dashboard setup in the future. I started working on a Cl to parse the gcov files and generate html as I couldn't get any other parser to work correctly. It stalled due to time, but it worked well enough as I tried to add coverage to the CSS parser code.

https://pdfium-review.googlesource.com/c/2411/

Patch-set: 2