Update patch set 1
Patch Set 1: Code-Review+1

The change looks good, but the description seems incorrect to me. The files are UTF-8, and the characters in them are representable in UTF-8. However VC++ doesn't *realize* that the files are UTF-8 so it compiles then in the local code page.

So, I'd say:

Tell VC++ that the pdfium source files are encoded in UTF-8 so that it doesn't interpret them as being encoded in the local code page.

Patch-set: 1
Reviewer: Gerrit User 5660 <5660@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Code-Review=+1