Update patch set 2
Patch Set 2:

Just to clarify, I don't necessarily mean to land this as is. I just meant it as a bit of a draft / suggestion to support Lei in working on the FreeType upgrade. Thanks for looking into it.

Patch-set: 2
Reviewer: Gerrit User 6715 <6715@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0