Update patch set 2
Patch Set 2:

For each of the added bits, can you make sure that there is a patch file that we can apply? Then in the future if we upgrade again we just have to apply the patches instead of finding the differences.

Patch-set: 2
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>