Update patch set 3
Patch Set 3:

So do we want to do this first, or move PDFium's FreeType to a DEPS-based checkout first?

In any case, for this to have a chance of working, we need to have an up to date build/ dir with the changes you have been making in the Chromium repo. I'm doing a roll here: https://pdfium-review.googlesource.com/c/3332/

Patch-set: 3
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>