Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> So do we want to do this first, or move PDFium's FreeType to a DEPS-based checkout first?
> 
> In any case, for this to have a chance of working, we need to have an up to date build/ dir with the changes you have been making in the Chromium repo. I'm doing a roll here: https://pdfium-review.googlesource.com/c/3332/

I'd prefer to have this first, so that we can move to shared compilation, but it looks like we need to have one change to the third_party/freetype BUILD.gn file to add  "-Wno-int-to-pointer-cast" to avoid a build failure on Clang. I'll add that. Then we could land this.

If the DEPS-based checkout CL is ready and green on the bots, we can it as well, but perhaps we should do one roll to Chromium in between.

Patch-set: 3
Reviewer: Gerrit User 6715 <6715@fca2d64a-0420-3532-b200-12848c95a47b>