Update patch set 3
Patch Set 3:

Dear Lei Zhang

As I mentioned in the previous review, it doesn't look to be necessary.

That being said, it was part of the changes featured in the zlib copy used by  PDFium.

There is no bug tracking the idea of upstreaming this change. Feel free to open one and reference it.

Also check:
https://pdfium-review.googlesource.com/c/3295

Patch-set: 3