Update patch set 9
Patch Set 9:

Having made all the changes locally, we're still hitting the cookie assert in the embeddertests like we used to.  

This may be in partitionalloc.   Here's a suggestion.

Write a brand-new unittest that simply allocates an 8163 byte buffer, and reallocs it up by an addtional 8163 bytes per iteration until it gets to like, dunno, 64 MB, and see if that trips the same assert.

Patch-set: 9
Reviewer: Gerrit User 5495 <5495@fca2d64a-0420-3532-b200-12848c95a47b>