Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Has this been tested in a Chrome build? Does everything still work correctly?
> 
> I am not quite sure how to test this as part of a Chrome build - can I trigger a Chrome bot against this CL?


The change needs to be pulled into a pdfium directory in a chromium checkout and then chrome built. We don't have any chrome bots on the PDFium waterfall. We could land and see if it blocks the chrome roll but that's a bit of a pain.

Patch-set: 1
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>