Update patch set 4
Patch Set 4:

So unfortunately this doesn't work when PDFium is built as part of Chromium. Unlike the jpeg target, there is no zlib target in Chromium's //third_party/BUILD.gn. In Chromium, if one wants to use system zlib, I believe the recommended route is to unbundle using the tools in build/linux/unbundle. I'm not sure we want to do that in PDFium, so let's see if we can come up with something else.

Patch-set: 4