Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> This is because we are doing memory intensive operations on things like bitmaps, right? I recently had to revert a std::vector CL in Chromium as well, because std::vector in debug mode can be really slow.


I didn't track down which of the fixedbufgrow items caused the slowdown but on mac/win you'd see a doubling of time in the bot runtimes when this was applied.  It's possible it's just one or two of them and we could go through and convert 1-by-1 but haven't gotten to it yet.

Patch-set: 2