Update patch set 4
Patch Set 4:

Do we understand the cause of the flakiness? Is it because the test files were missing the end keywords? I noticed the tests were running against both bug_665467.in and bug_665467.pdf. Could that be part of the cause?

Patch-set: 4
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>