Update patch set 1
Patch Set 1:

> This means all embedders have to fix their FPDFAvail_Destroy() 
> + FPDF_CloseDocument() call ordering, including in Chromium's pdf/pdfium/pdfium_engine.cc and pdf/pdfium/fuzzers/pdfium_fuzzer_helper.cc.
> 
> It also feels unintuitive, because FPDFAvail_GetDocument() can create a FPDF_DOCUMENT using a FPDF_AVAIL. So if the FPDF_AVAIL came first, one would expect it to be destroyed last.

Ah, which makes the linkage between the avail and the document dubious.
Recall the document is owned by the embedder across the C API.
The same is true for the avail.  So this is suggesting maybe that the doc
needs to be passed into the avail calls that need it?  Hmmm.

Patch-set: 1
Reviewer: Gerrit User 5495 <5495@fca2d64a-0420-3532-b200-12848c95a47b>