Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I worry about using checked numerics in pixel manipulation routines, which need to be fast.  Are we sure we don't degrade performance?  Alternatively, can we change to types (unsigned, floats) where overflow is either defined or avoided?

This is not a pixel manipulation routine, though. The method is used to calculate the color components of a path in a shading pattern. I ran release pdfium_test on the 3 corpus tests affected by the change and this CL did not seem to change the run time significantly. I'm not sure we can use unsigned, and using float would probably result in worse inefficiency.

Patch-set: 3