Update patch set 1
Patch Set 1:

Interesting, so I think this: the generate method is supposed to be used when we want to reload the page or save it after changing the page, since those methods probably ask the page for its Contents stream, which is unchanged when objects are inserted. The problem is that this method does not know which objects have been changed and which haven't. So it tries to re-create the whole Contents of the page, but it's currently doing a bad job at it (we are missing support for a ton of things).

Your solution is not a good one, though. First, you need to use the page object list, not the 'over' one you created, so that when you call render the stuff actually renders. So these new methods would be unnecessary. And second, you will end up with duplicates of objects because you're adding to the current Contents all the ones that the generate method successfully parses.

Patch-set: 1
CC: Gerrit User 6295 <6295@fca2d64a-0420-3532-b200-12848c95a47b>