Update patch set 1
Patch Set 1: Commit-Queue+1

(21 comments)

From a high level, this is looking good so far. I had a couple suggestions on changing the APIs and would like to see this integrated into pdfium_test as a flag to dump the annotations to console, but this looks like the right direction to me so far.

Comments inline (a bunch of them are nits).

Patch-set: 1
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=+1