Update patch set 4
Patch Set 4:

So, previously the clip region was shared, so if we intersected it or did any operations on it, it would reflect in both places. That's no longer the case, does that have any implications for the code?

Patch-set: 4
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>