Update patch set 8
Patch Set 8:

(12 comments)

This is looking a lot better. The code feels a lot more natural in that we just figure out if we have to regenerate or not instead of depending on the user setting a flag. Thanks for iterating on this, I think we're converging on a really good solution.

Patch-set: 8
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>