Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > (1 comment)
> 
> I couldn't anything on the specs about that. I should note "http://www.google.co.kr/)0" is a valid URL according to RFC 1738. Wei Li, do you know if this is spec'ed somewhere?
> 
> If there's nothing on the specs, this is done on a best-effort basis, so we could apply these parentheses stripping rules, then I'd update this test case and leave it commented out, then implement the rules in a following CL.
> 
> If there is something on the spec, we'll know what to do.


google.co.kr/)0 is valid, but we know given the input string that it's probably wrong. I think we're safe to assume that if the input starts with a ( we shouldn't include the matching ) in the URL as it isn't part of the URL. (If the open ( is in the URL then we shouldn't strip the closing ), just in the case of things like (http://example.com) do we know that the ()'s can be stripped.

Patch-set: 6
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>