Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> What was the impetus for this? Seems like we're just moving the comparitor further from the place of comparison?

In some of the other cases of applying this pattern, it was commented that it made the code overly nested/dense to have the comparator defined in the args, thus hard to read. I am just tweak it here for consistency with the other usages.

Patch-set: 2
Reviewer: Gerrit User 7065 <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0