Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> Is it possible to add a test for this?

What would the test do?

The change mimics the code in fx_agg_device.cpp so that if exactly the same rectangle is returned for rectangular clips. While I don't know if this is
fixing a bug, it seems worthwhile to exactly match existing behavior where possible to eliminate unnecessary rabbit holes. Not sure how you would like to test that.

Patch-set: 1
CC: Gerrit User 6215 <6215@fca2d64a-0420-3532-b200-12848c95a47b>