Update patch set 2
Patch Set 2:

I don't think this CL is right. We should be able to get away with just releasing the UnownedPtr instead of changing to another type. This isn't a double free, the OnownedPtr Probe check just makes sure that the thing the UnownedPtr points too lives longer then the UnownedPtr itself which is currently being violated because we Unref, which destroys the object and then the destructor cleans up the UnownedPtr. Releasing the UnownedPtr then Unref'ing should fix the issue.

Patch-set: 2