Update patch set 2
Patch Set 2:

Hi Dan, here are some starter APIs and test for extracting embedded attachments.

The idea is that a CPDF_NameTree can be used to hold the embedded files and their names for a document. For each embedded file, a wrapper CPDF_FileSpec can be created, which will have helper functions to get the file's platform-dependent name, data, and other parameters. 

(During document parsing, the parser doesn't actually skip any objects, even if they are not supported; so we don't need to add any code in the parser to parse the embedded file objects.)

I designed the APIs such that CPDF_NameTree is created every time any file info is requested. NameTree is basically just a wrapper class, so this wouldn't be much more work. This is the same flow as FPDF_CountNamedDests(),  FPDF_GetNamedDest(), etc.

Alternatively, we can return CPDF_NameTree to the user to avoid creating it every time, but that adds an extra step in the flow.

Another alternative is to have CPDF_Document create its name tree and save it as a member variable during parsing, but I don't see the point of doing this yet - most of the document objects like page and annot are created on demand, there doesn't seem to be a need to save them permanently. 

Let me know what you think of this design!

Patch-set: 2
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>
Reviewer: Gerrit User 6975 <6975@fca2d64a-0420-3532-b200-12848c95a47b>