Update patch set 3
Patch Set 3:

This is better, but I don't think it's quite right. I think what we need to do is just re-write the format strings. Looking at the documentation this doesn't handle parts of the input.

Each conversion specification is constructed as follows:
  %[,nDecSep][cFlags][nWidth][.nPrecision]cConvChar

nDecSep A comma character (,) followed by a digit that indicates the decimal/separator format:
  0 — Comma separated, period decimal point
  1 — No separator, period decimal point
  2 — Period separated, comma decimal point
  3 — No separator, comma decimal point

cFlags Only valid for numeric conversions and consists of a number of characters (in any order), which will modify the specification:
    + — Specifies that the number will always be formatted with a sign.
space — If the first character is not a sign, a space will be prefixed.
    0 — Specifies padding to the field with leading zeros.
    # — Specifies an alternate output form. For f, the output will always
        have a decimal point.

nWidth A number specifying a minimum field width. The converted argument is formatted to be at least this many characters wide, including the sign and
decimal point, and may be wider if necessary. If the converted argument has
fewer characters than the field width, it is padded on the left to make up the
field width. The padding character is normally a space, but is 0 if the zero
padding flag is present (cFlags contains 0).

nPrecision A period character (.) followed by a number that specifies the number of digits afterthe decimal point for float conversions.

cConvChar Indicates how the argument should be interpreted:
  d — Integer (truncating if necessary)
  f — Floating-point number
  s — String
  x — Integer (truncating if necessary) and formatted in unsigned
      hexadecimal
notation

Patch-set: 3
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>