Update patch set 3
Patch Set 3:

(4 comments)

Hi Lei, Dan and Jane,

In this patchset, I have moved FORM_GetAnnot() to FPDFAnnot_GetInteractiveFormAnnotAtPoint(), as well as the associated embedder tests. Additionally, I simplified some of the checks in this method, and made it more clear in the user doc comment that "point" refers to a point on a page contained in an annotation rectangle.

Thanks for reviewing.

Diana

Patch-set: 3
Reviewer: Gerrit User 7040 <7040@fca2d64a-0420-3532-b200-12848c95a47b>