Update patch set 13
Patch Set 13:

> Patch Set 12:
> 
> > Patch Set 12:
> > 
> > There seems to be a bunch of stuff in this CL, can we break this apart into smaller CLs to make review easier?
> 
> Unfortunately no.
> I split the very tied logics (like knot). (Because, its are very difficult to understand and update).
> 
> If I refactor only decryption / encryption, the parsing will be failed.
> If I refactor only parsing, the serialization will be failed.
> If I refactor only serialization, the decryption / encrytiption will be failed.
> Etc.
> 
> I can refactor this only by one block.

There are some steps that can be done in preparation for this change:
- Add CPDF_ObjectWalker
- Refactor CPDF_Stream::ReplaceData out of SetData
- Add the return value to CPDF_Dictionary::RemoveFor
- Add the new methods to CPDF_CryptoHandler

I know that if you had uploaded these changes separately, no one would understand why they are necessary. Now that we know what the end result will be like (and it looks like a great improvement to me), these pieces can be broken out and you can slim this CL down to the most complex changes. It'll be much clearer to reviewers.

Patch-set: 13
CC: Gerrit User 7035 <7035@fca2d64a-0420-3532-b200-12848c95a47b>