Update patch set 13
Patch Set 13:

(11 comments)

I think my one big question with this CL is, this changes the encryption/decryption to all be handled at the end by a global tree walk. That seems like it could have significant performance issues for large documents.

Would we be able to consolidate the decryption/encryption logic in the crypto handler, but just pass in the streams of data we nee to decrypt/encrypt as we do the parsing? Why does this need to be done as a  single tree walk at the end?

Patch-set: 13
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>