Update patch set 10
Patch Set 10: Commit-Queue+2

> Patch Set 10:
> 
> > Patch Set 10:
> > 
> > > Patch Set 10:
> > > 
> > > > Patch Set 10: Code-Review+1
> > > 
> > > I just see only "+1" button in the Commit-Queue line. :(
> > 
> > If you refresh, how about now?
> No, may be i'm looking in incorrect place?
> https://yadi.sk/i/2yqwU4Ce3LSRJr


That's the right spot, maybe there is a delay on the flag getting set? You were in the group with CQ access but not commit access which is what I changed. I'll +2 this and we'll see if it corrects itself with a bit of time.

Patch-set: 10
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=+2