Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > (3 comments)
> > 
> > > Patch Set 2:
> > > 
> > > (3 comments)
> > > 
> > > Is the long term plan to use CFX_FloatRect for everything? i.e. get rid CFX_Rect and CFX_RectF as well?
> > 
> > 
> > Maybe? CFX_Rect and CFX_RectF have a flipped y-axis to CFX_FloatRect. If we want to keep an int32_t rectangle then maybe it would be better to keep CFX_Rect and CFX_RectF as they share the same base code and fold CFX_FloatRect into those if possible?
> 
> I suspect you may not be able to get rid of FX_RECT entirely. PDF user space has y-axis runs upwards and so does TrueType font metrics. (Yet somehow font code uses FX_RECT quite a bit.) Whereas the rendering code, which is probably inspired by GDI, has y-axis running downwards. (Same goes for Skia, and presumably AGG.)


Well, the other option is to replace FX_RECT with CFX_Rect which has the y-axis running down. Although it would be nice to just have one set of rectangles and we flip things on input/output if we really need too. Having to know which type of rect in which place is pretty confusing.

I guess the question we should ask is, which way do we want our y-axis to run. My vote would be to keep the pdf y-axis running up (CFX_FloatRect) as we're a PDF library and flip the incoming truetype glyphs and the output to the graphics engines.

Patch-set: 3
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>