Update patch set 14
Patch Set 14:

(4 comments)

I used this to compare how enabling Skia paths changes rendering. It can be used as a pixel test that runs on any set of pdfs.

It bothers me a bit that it's not in the original conception of safetynet_compare, but the options and code would be so similar if this was a separate script that I opted to keep it in the same one and let it do both things, keeping the image comparison code in a separate module.

The next logical step might be refactor the performance stuff into a module too and have the compare script be a more generic thing that only deals with checking out versions, etc.

Patch-set: 14
Reviewer: Gerrit User 7035 <7035@fca2d64a-0420-3532-b200-12848c95a47b>