Update patch set 4
Patch Set 4:

Cary, from a graphics perspective, is it better for us to have all our rectangles going in a uniform direction and then doing the flip in the Skia code, or would it be better for us to have two rectangle types with opposite y-axis that we use as needed?

Or, does it really not matter and we should just pick something and stick with it?

Patch-set: 4
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>