Update patch set 3
Patch Set 3: Commit-Queue+1

(1 comment)

> Patch Set 3:
> 
> > Patch Set 2:
> > 
> > (1 comment)
> 
> Coverage is inherently about testing, so IMO it makes sense to be moved to testing/tools.

After thinking about this a bit, I agree that testing and coverage scripts should be in the same place. I think it is a bit odd that we have testing/tools instead of tools/testing, but I suspect is more of a subtle personal preference then a compelling argument for script locations. The coverage scripts should probably be moved into testing/tools.

If we are going to move where the coverage scripts live, I would suggest doing it in a separate CL, so that there is a single change with the file movements and no functional changes, so that patching and merging are simpler. It can be either before or after this one, I don't feel strongly about it.

Patch-set: 3
Reviewer: Gerrit User 7065 <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=+1