Update patch set 8
Patch Set 8:

This is .... hard to follow. Can we do the folder move in one CL and then the patch renumber in another and a third to do the upgrade? It's hard to figure out what went where at the moment.

Patch-set: 8
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>