Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> As I look at this, I'm not 100% sure this does the same thing.  Let's go back to the previous PS.  Sorry bout that.


I believe this is the same, the points are always identical, so we don't add the MoveTo calls, we have to force the path closed but the old way did that already. This way is a lot clearer on what it's doing by just adding lines I think.

Patch-set: 6
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0