Update patch set 11
Patch Set 11:

> > Felix - Have you tried building with both is_official_build = true and is_component_build = true? I was not able to get it to build. See https://bugs.chromium.org/p/pdfium/issues/detail?id=855
> 
> I don't think I ever succeeded doing a combination of a shared library build and is_official_build with gcc on Linux, maybe with Clang, but I am pretty sure it used to work on Windows.

Official component builds are unsupported in Chromium. This makes intuitive sense to me because official builds are supposed to be maximum speed and component builds are for ease of development with compromised speed. They are fundamentally at odds. There is an assert to prohibit this combination:

https://cs.chromium.org/chromium/src/BUILD.gn?q=build.gn&sq=package:chromium&dr&l=40

Patch-set: 11
Reviewer: Gerrit User 5660 <5660@fca2d64a-0420-3532-b200-12848c95a47b>