Update patch set 3
Patch Set 3: Commit-Queue+1

(3 comments)

Played around with const-ness a bunch to get an understanding of what the compiler is doing. If both the const and non-const versions of the [] operator are available, then the compiler is using the non-const version, since it doesn't appear to be contextually able to determine it is safe to use the const version, even for simple cases like const wchar_t foo = str[0]; The "fix" for this is to have the call sites use const versions of the string objects, either through changing the declaration or using static_cast to coerce at call time.

Doing this I think would be strictly worse then the current implementation, since it would require developers to be careful about const everywhere to avoid a performance hit when reading an element.

Talking with Dan, we settled on the idea of removing GetAt, implementing const [], and retaining SetAt. This way reading using [] is guaranteed to behave intuitively, and mutating the string requires an explicitly call. There is no possibility of accidentally causing Realloc to be called by reading.

Patch-set: 3
Reviewer: Gerrit User 7065 <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=+1