Update patch set 6
Patch Set 6: Commit-Queue+1

(6 comments)

> Patch Set 2:
> 
> can we break this into smaller pieces? It's making my head hurt :) ...


I've broken it up a bit. Renaming CFDE_TxtEdtEngine to CFDE_TextEditEngine makes git detect it as a new file, which it basically is. The new EditEngine is a complete rewrite as a gap buffer editor. The undo/redo system was a complete rewrite as a ring buffer.

There are bits of the code which are similar to the old code, mostly around using the text break code.

Patch-set: 6
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=+1