Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> So is step 5 where the big allocation fails, or is it step 6 ?

In production: step 5 alloc succeeds, then on 6 the bitmap alloc is determined to be too large, so we fail the decode, so the image won't appear in the doc.

On the fuzzers: step 5 succeeds, we immediately get killed for over allocating and a report is generated.

Patch-set: 1
Reviewer: Gerrit User 7065 <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0