Update patch set 2
Abandoned

My understanding of the problem was incorrect. I thought that decode would fail at a later stage due to us being unwilling to allocate a large enough bitmap to receive the JPX image, so there was a mismatch between the memory limits in PDFium and OpenJPEG. Digging in more, the bitmap limit is the size of a valid alloc, so if it is going to fail, the JPX decode will fail earlier. If you give the fuzzer enough RAM to work with, it will eventually successfully decode the image.

Patch-set: 2
Status: abandoned
Tag: autogenerated:gerrit:abandon