Update patch set 5
Patch Set 5: Commit-Queue+1

> Patch Set 4:
> 
> When is the next branch?  We may be able to make our lives easier when it comes to merging patches if we align this immediately before/after a branch is taken, though I'm not convinced which is better or if it matters.  Thoughts?

Next branch is supposed to be Oct 12. I am preferential to landing this before the branch, because then fixes on HEAD should avoid merge conflicts. If we wait until after the branch, all the fixes on HEAD may need to be rebased when cherry picking them to the release branch, since this change will be on HEAD, but not the branch.

This of course is predicated on me thinking this is a relatively low risk change at this point, since the underlying type change has been sitting on HEAD for awhile without needing to be reverted.

Patch-set: 5
Reviewer: Gerrit User 7065 <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=+1