Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > Ping?
> > > 
> > > Can you file a bug about this? I can see the embeddertest, but I'm not sure how this happens in practice?
> > 
> > I have not seen this in practice. And i think this is very rare case.
> > I have found this bug using FakeFileAccess (see https://pdfium-review.googlesource.com/c/pdfium/+/11890/7), because with it the exists test FPDFViewEmbeddertest.LinearizedDocument was failed.
> > 
> > Since the FakeFileAccess and TestAsyncLoader is simulation of PDFiumEngine's logic, i sure, it happens in practice.
> 
> 
> That sounds like it's a bug with the FakeFileAccess then since it doesn't fail without that change?

Oo
Again:
The CPDF_DataAvail request data from 900 to file end
After data reseived the CPDF_Parser try to read data using CPDF_SyntaxParser from 900 to file end.
But the CPDF_SyntaxParser(original) read data from 512(not 900) to file end to read char at 900 offset.

The FakeFileAccess have detected that we trying to read NOT REQUESTED data, and do not allow read it. As result exist test was failed.

Are you argee, that when the requested range is from 900 to file end, will be INCORRECT to read range from 512(not 900) to file end. for read char at 900 position?

BTW: The test was not failed before, because it is never do test on unavailable data (but should).
More precisely, there are no tests in the pdfium, which are testing cases when data is unavailable except two my. As result all embedder test are fake, because it is working only in case, when whole file available. The FakeFileAccess will fix this, and make the embedder tests not fake.

Patch-set: 1
Reviewer: Gerrit User 7160 <7160@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0