Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > I still think it will be better for us in the long run to figure out how these classes are related and if we can break some of the dependencies between them. I'm going to take a look at this starting next week to see what I can figure out.
> > 
> > I want to ask you again to use this CL as a temporary solution.
> > 
> > This will allow us to rework the CPDF_DataAvail in order to simplify and split the components from it. That in turn will allow to create normal fix for document owning issue after.
> > 
> > As you can see in https://pdfium-review.googlesource.com/c/pdfium/+/13530 
> > Much of code and some members is removed from CPDF_DataAvail, and uploading of that CL is making start splitting logic of CPDF_DataAvail more simple, as you are wishing.
> 
> 
> I don't think this CL is the right solution. 

Again, this is not right solution, this is TEMPORARY solution.

>It makes the code harder to understand (adding more #ifdef PDF_ENABLE_XFA blocks) along with having retained pointers which Leak().

This is only in one place, which will be restored after right solution.

> 
> I think we need to look at how we can break document out of CPDF_DataAvail instead of changing the document ownership. 

No good way to do this at this moment.
Before move the document up to the FPDF_AvailContext (or in another place), we should rework CPDF_DataAvail.

As result we are in vicious circle. 
1) We can not move the document and change it ownership because CPDF_DataAvail is not ready for this.(2)
2) We can not make CPDF_DataAvail to ready for move document and change it ownership because ownership of document in original code already incorrect. (3)
3) We can not fix ownership of document  because we can not move the document and change it ownership (1).

I want to temporary fix (3), to break this vicious circle, then solve (2) , then solve (1) , and then solve (3) again, but with right fix.

Also moving the document up to the FPDF_AvailContext will not fix original problem, because FPDF_AvailContext should outlive the document also how CPDF_DataAvail and this is same issue.

> If we move the document up to the FPDF_AvailContext can we then place the PageAvail object in that context as well and remove the need to plumb this through DataAvail?

Yes, if we move, but we CAN NOT MOVE document at this moment.

Patch-set: 3
Reviewer: Gerrit User 7160 <7160@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0