Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> The only problem is that this makes it easier to upload expected pngs without first looking at them. But I guess Polygerrit helps with side-by-side diffs.

I was using Polygerrit to look at the diffs anyways. Would you prefer that I write the new images to a temp dir, add a script that generates the comparison in an HTML (I already have this in https://pdfium-review.googlesource.com/c/pdfium/+/8830) and open the HTML? That might be a better approach, although it's much more code to add than this. It would also have the added benefit of making it easy to aggregate results from several OS'es (which is something I'd have to follow up with on this CL).

Patch-set: 4
Reviewer: Gerrit User 7035 <7035@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0