Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I don't think this does what you think it does, in that the multiply takes place first in an unsafe manner before being promoted and checked as a safe type.
> 
> +jschuh just because my memory is fuzzy.

Yes, it makes sense that that should be the case when there is nothing SAFE on the right side of the "=". Win compiler is also complaining about it so I have to split it up anyways.

Patch-set: 1
Reviewer: Gerrit User 6295 <6295@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0