Update patch set 1
Patch Set 1:

This is making the assumption the CJS_* object is the primary class and not the subclass of the EmbedObj. I'd like to hide the EmbedObj but we do access them in various places which makes it more difficult.

Thoughts?

Patch-set: 1
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>