Update patch set 3
Patch Set 3:

So, for other embedders. The loading of the persisted file would work if the specifically named file exists. But, currently, nothing new would ever get written back to it. There is no existing call to CommitGlobalPersisitentVariables the persisted data, so we never add any new entries to be written out.

So, if you start with an empty file, it will always be empty. If you have a file it will no longer be read, but there was no way to write it in the first place.

Patch-set: 3
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>