Update patch set 5
Patch Set 5: Commit-Queue+1

The original test is for the initial feature fix, where we were not removing the soft hyphen from the outputted string, which was fixed by c5a729ec5ada8ef0edc3f4d9976e96865f516598.  The commit that you are referencing was a follow on for a specific issue that the fuzzers running under ASAN caught. I have added a new test for that issue. If you revert the commit you are referencing and turn on ASAN it will hit an issue when running the bug_782596 test. If you revery both CLs, regardless of ASAN, the GetTextWithHyphen test should fail.

Patch-set: 5
Label: Commit-Queue=+1