Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > Can you pls add a small test too?
> > 
> > I don't know how to add a small test for this, as there is no coverage on that file to build on.
> > 
> > My first idea is modifying test_runner.py to make sure .ps output hasn't changed, but we are trying to get rid of our gold files for corpus and pixel tests and moving to Skia Gold, which probably doesn't support .ps. I guess I could write a test that mostly copies what pdfium_test does with --ps2 and --ps3. What do you suggest?
> 
> I am not sure about Skia Gold move. For current tests, maybe adding a flag in test_runner, and a corpus test for Windows only. You may want use a text only pdf file to demo this bug.

I'm trying to get access to Dan's Windows machine to write the test. However it's a bad enough bug that we don't want to leave it hanging for days. Can we do the test in a follow up so the fix gets in? It's a trivial fix, too.

Patch-set: 3
Reviewer: Gerrit User 7035 <7035@fca2d64a-0420-3532-b200-12848c95a47b>