Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Should we really be moving more code into the pdfium namespace, and then dropping pdfium:: for Optional, since we don't need pdfium:: inside the pdfium namespace?
> 
> In the end it doesn't matter, since Optional will be replaced with std::optional.

Moving more stuff into the pdfium:: namespace is a good goal but it will take a long time to get there and this change makes a bunch of code easier to read. In theory all this code should be in the pdfium:: namespace so once we get there we can just remove the using command.

Patch-set: 2