Move remaining jpeg error handling up a level
This moves the setjmps needed for handling fatal errors in the jpeg
library up a level to be in line with how other instances of this are
being modified. This additionally reduces the number of times that
setjmp needs to be called and documents why it is occurring.

This covers the Start and ReadScanLine methods. It also adds in
setting the error member, which had been missed in previous CLs.

BUG=pdfium:986

Change-Id: I7db87288ffe0ee8b29899d97035c30ad812da76a