Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I don't know if we should really remove FF_Draw. The class hierarchy is correct with it, and looks off without it. Looking at the readme it makes sense to have this separation between FFDraws and FFFields.
> 
> Well, the class hierarchy is correct either way, it just changes how deep it is. I'd rather not keep the class just so the README looks nice. We could always just add a comment in the README that these are Visual vs Editable classes or something.

Yeah, let's leave the class in the readme with some annotation that FFDraw doesn't actually exist anymore and remove it from the code, how about that?

It's not about it looking nice, it's the fact that the documentation makes more sense with it, meaning the correct structure might be the previous one.

Patch-set: 1
Reviewer: Gerrit User 7035 <7035@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Code-Review=+1