[JS] Move CFXJS_Engine from v8::Context to v8::Isolate.
Currently we store the CFXJS_Engine into the aligned embedder data slot
of the v8::Context we create when initializing the CFXJS_Engine. We then
retrieve the engine by accessing the aligned embedder data slot of the
current context. This works fine in non-XFA where we have one context
but in XFA we have many many contexts. There is no guarantee that when
we pull the context off the top of the context stack it will have the
pointer to our engine.

This CL moves the engine pointer out of the aligned data slot and puts
it into the SetData of the v8::Isolate. We can then access the engine
data out of the isolate itself instead of going through the aligned
embedder data of the context.

BUG: chromium:773229
Change-Id: Ia7b05cc05f4c6770f91d96b530390afbf95ffb33