Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 5:
> > 
> > (1 comment)
> > 
> > Lei, what do you think about changing the public api for this function to return signed?
> 
> I think new API + deprecating existing API is cleaner. If there is a new API, it is way easier to check if PDFium callers have fixed things on their side.

The thing is that there already is a case in which -1 is returned. There's even a test making sure of that. Should we then fix the current API to always return 0 on errors, as well as create one that returns -1?

Patch-set: 7
Reviewer: Gerrit User 7035 <7035@fca2d64a-0420-3532-b200-12848c95a47b>