[jumbo] Compile fix: Move a GetPageContent to a more limited scope
There is a GetPageContent in pdfsdk/fpdfppo.cpp and one in
fpdfsdk/fpdf_transformpage.cpp, both in the anonymous namespace. In jumbo
builds those two files compile in the same translation unit and then
they share the same anonymous namespace and clash.

Given that the function is small and that I saw no natural place to
share it, this patch moves one instance into the class that is
using it.

Change-Id: Icbdf69bd83887c6e5cf31d2080a5b12b5710691e