Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1 Commit-Queue+2
> 
> Any idea why this didn't get caught by our Jumbo trybot?

I don't know for sure but it might be that you are building with smaller jumbo units than us non-goma people (and the Jumbo Linux fyi builder). With goma the default jumbo size is just 8 files at a time. Otherwise it's 50. So if the colliding files are ~10-20 lines apart in the sources list, goma builders will never build them together but non-goma builders are likely to do so.

If so, the good news is that goma users won't have noticed this. Bad news is that it isn't Linux only as I first assumed.

Patch-set: 2
Reviewer: Gerrit User 7314 <7314@fca2d64a-0420-3532-b200-12848c95a47b>