Remove V8 context from CJS_V8
Makes CFXJS_Engine and CFXJSE_Engine consistent with each other in
that both have a V8 context to themselves, and not one inheritted
from the CJS_V8 (which is now more per-isolate than per-context).

Consolidate NewLocalContext() and GetPersistentContext(), which both
did the exact same thing under the covers once inside v8 land.

Rename a few things to make it simpler.

Change-Id: I68905db9ad44253063da235fcb276a75627a2dbc