Update patch set 4
Patch Set 4:

> Patch Set 4:
> It's b.  From my understanding of the existing rendering tests which compared against a pre-computed hash value.  How do I compute the correct hash value before getting the correct image?

Run the test locally and see what the hash value is from the rendered bitmap. Then run it through the trybots and see if they agree.

If you are confident this is the right fix for the problem you observed, then that's fine. Just keep in mind since there is a TODO to add rendering tests for N-up, it would be good to provide test coverage for this case.

Patch-set: 4
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>