Use temporary iterator to avoid potential OOB
In the existing code pCharPos is manipulated directly without being
reset. This means that for the second iteration it is at the end of
the range instead of the start. This CL introduces temporary iterators
that are intialized to the value of pCharPos and then manipulated to
avoid this issue and having to reset pCharPos.

BUG=chromium:648177

Change-Id: I5c9344c1b67a015b01470a0dc337361552ffd447